imporove UI in messenge board
