+++ src/components/MessengeBoard.vue
@@ -9,27 +9,27 @@
         公開聊天室
       </div>
     </div>
-    <div class="board-wrapper">
+    <div id="board-wrapper" class="board-wrapper" @click="scrollToBottom">
       <div class="messages">
         <div v-for="(message, index) in messages" :key="`msg-${index}`">
-          <div class="broacast-message-wrapper" v-if="message.type === 'userComein'">
+          <div class="broacast-message-wrapper" v-if="message.type === 'userComein' && currentUser.id !== message.UserId">
             <div class="broacast-message">{{ message.message }}</div>
           </div>
-          <div class="other message" v-if="message.type !== 'userComein' && currentUser.id !== message.id">
-            <div class="avatar" :style="{ background: `url(${message.avatar}) no-repeat center/cover` }"></div>
+          <div class="other message" v-if="message.type !== 'userComein' && currentUser.id !== message.UserId">
+            <div class="avatar" :style="{ background: `url(${message.User ? message.User.avatar : message.avatar}) no-repeat center/cover` }"></div>
             <div class="wrapper">
               <div class="text">{{ message.message }}</div>
               <div class="time">{{ message.createdAt | fromNow }}</div>
             </div>
           </div>
-          <div class="self message" v-if="message.type !== 'userComein' && currentUser.id === message.id">
+          <div class="self message" v-if="message.type !== 'userComein' && currentUser.id === message.UserId">
             <div class="text">{{ message.message }}</div>
             <div class="time">{{ message.createdAt | fromNow }}</div>
           </div>
         </div>
       </div>
     </div>
-    <form @submit.prevent="sendMessage">
+    <form @submit.prevent="sendMessage" @click="scrollToBottom">
       <div class="text-wrapper">
         <input v-model="message" class="text" placeholder="輸入訊息..." onfocus="this.placeholder = ''" onblur="this.placeholder = '輸入訊息...'" />
         <div class="icon-wrapper" type="submit" @click="sendMessage">
@@ -63,38 +63,56 @@ export default {
     }
   },
   mounted() {
+    this.fetchChatroom()
     this.$socket.emit('chatting', { ...this.currentUser, createdAt: new Date(), type: 'userComein' })
     this.$socket.on('userOnline', (data) => {
       this.$emit('someoneComein', data)
+      const scroll = this.$el.querySelector('#board-wrapper')
+      scroll.scrollTop = scroll.scrollHeight
+      scroll.animate({scrollTop: scroll.scrollHeight})
     })
     this.$socket.on('msg', (data) => {
-      this.messages = [...this.messages, {id: data.id, avatar: data.avatar, message: data.message, createdAt: data.createdAt, type: data.type}]
+      this.messages = [...this.messages, {UserId: data.UserId, avatar: data.avatar, message: data.message, createdAt: data.createdAt, type: data.type, User: data.User}]
       this.messages.sort((a, b) => {
         return a.createdAt > b.createdAt ? 1 : -1;
       })
+      const scroll = this.$el.querySelector('#board-wrapper')
+      scroll.scrollTop = scroll.scrollHeight
+      scroll.animate({scrollTop: scroll.scrollHeight})
     }),
     this.$socket.on('newclientlogin', (data) => {
-      this.messages = [...this.messages, {id: data.id, avatar: data.avatar, message: data.message, createdAt: data.createdAt, type: data.type}]
+      this.messages = [...this.messages, {UserId: data.UserId, avatar: data.avatar, message: data.message, createdAt: data.createdAt, type: data.type, User: data.User}]
       this.messages.sort((a, b) => {
         return a.createdAt > b.createdAt ? 1 : -1;
       })
+      const scroll = this.$el.querySelector('#board-wrapper')
+      scroll.scrollTop = scroll.scrollHeight
+      scroll.animate({scrollTop: scroll.scrollHeight})
     })
   },
   beforeDestroy () {
-    this.leaveChatroom ()
+    this.leaveChatroom()
     this.$socket.emit('leave', this.currentUser.id)
   },
   watch: {
     messages () {
-      const scroll = document.querySelector('.board-wrapper')
+      const scroll = this.$el.querySelector('#board-wrapper')
       scroll.scrollTop = scroll.scrollHeight
       scroll.animate({scrollTop: scroll.scrollHeight})
     }
+  },
+  updated() {
+    this.scrollToBottom()
   },
   computed: {
     ...mapState(['currentUser', 'isAuthenticated'])
   },
   methods: {
+    scrollToBottom () {
+      const scroll = this.$el.querySelector('#board-wrapper')
+      scroll.scrollTop = scroll.scrollHeight
+      scroll.animate({scrollTop: scroll.scrollHeight})
+    },
     sendMessage(e) {
       e.preventDefault()
       if (this.message === '') {
@@ -103,9 +121,9 @@ export default {
           title: '請輸入訊息'
         })
       }
-      this.$socket.emit('send message', {id: this.currentUser.id, avatar: this.currentUser.avatar, message: this.message, createdAt: new Date(), type: 'chat'})
+      this.$socket.emit('send message', {UserId: this.currentUser.id, avatar: this.currentUser.avatar, message: this.message, createdAt: new Date(), type: 'chat', User: this.currentUser})
       this.message = ''
-      const scroll = document.querySelector('.board-wrapper')
+      const scroll = this.$el.querySelector('#board-wrapper')
       scroll.scrollTop = scroll.scrollHeight
       scroll.animate({scrollTop: scroll.scrollHeight})
     },
@@ -117,6 +135,30 @@ export default {
         }
       } catch (error) {
         console.log(error)
+      }
+    },
+    async fetchChatroom () {
+      const loader = this.$loading.show({
+        isFullPage: true,
+        opacity: 1
+      }, { default: this.$createElement('MyLoading') })
+      try {
+        const { data } = await chatAPI.getChatRoom()
+        this.messages = data.histroy.map(d => ({
+          ...d,
+          type: d.message.indexOf('上線') > 0 ? 'userComein' : 'chat',
+        }))
+        const scroll = this.$el.querySelector('#board-wrapper')
+        scroll.scrollTop = scroll.scrollHeight
+        scroll.animate({scrollTop: scroll.scrollHeight})
+        loader.hide()
+      } catch (error) {
+        loader.hide()
+        console.log(error)
+        Toast.fire({
+          icon: 'error',
+          title: '目前無法連線聊天室，請稍候'
+        })
       }
     },
   }
