+++ src/utils/helpers.js
@@ -1,8 +1,8 @@
 import axios from 'axios'
 import Swal from 'sweetalert2'
 
-const proxyURL = 'https://cors-anywhere.herokuapp.com/'
-const baseURL = proxyURL + 'https://simple-twitter-demo-ac.herokuapp.com/api'
+// const proxyURL = 'https://cors-anywhere.herokuapp.com/'
+const baseURL = 'http://localhost:3000/api' // proxyURL + 'https://simple-twitter-demo-ac.herokuapp.com/api'
 
 export const apiHelper = axios.create({
   baseURL
